+++ index.html
@@ -6,7 +6,6 @@
   <link rel="stylesheet" href="css/normalize.css">
   <link rel="stylesheet" href="css/style.css">
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
-  <script src="/js/functions.js"></script>
   <title>Кекстаграм</title>
 </head>
 
@@ -228,5 +227,6 @@
     </section>
   </template>
 
+<script src="/js/main.js" type="module"></script>
 </body>
 </html>
