+++ js/functions.js
@@ -3,9 +3,9 @@
 const checkStringLength = (string, length) => string.length <= length;
 
 // Тесты для проверки длины строки
-console.log(checkStringLength('проверяемая строка', 20));
-console.log(checkStringLength('проверяемая строка', 18));
-console.log(checkStringLength('проверяемая строка', 10));
+checkStringLength('проверяемая строка', 20);
+checkStringLength('проверяемая строка', 18);
+checkStringLength('проверяемая строка', 10);
 
 // Функция для проверки, является ли строка палиндромом
 
@@ -15,7 +15,7 @@ const checkStringPalindrome = (string) => {
 };
 
 // Тесты для проверки, является ли строка палиндромом
-console.log(checkStringPalindrome('Лёша на полке клопа нашёл '));
+checkStringPalindrome('Лёша на полке клопа нашёл ');
 
 // Функция принимает строку, извлекает содержащиеся в ней цифры от 0 до 9 и возвращает их в виде целого положительного числа.
 // Если в строке нет ни одной цифры, функция должна вернуть NaN
@@ -26,4 +26,4 @@ const findNumber = (string) => {
 };
 
 // Тесты для проверки извлечения цифр
-console.log(findNumber(555464161));
+findNumber(555464161);
